Moved domain to shop directory part 3
