+++ common/bootstrap/SetUp.php
@@ -2,7 +2,7 @@
 
 namespace common\bootstrap;
 
-use frontend\services\contact\ContactService;
+use shop\services\ContactService;
 use yii\base\BootstrapInterface;
 use yii\mail\MailerInterface;
