--- frontend/tests/unit/forms/ResetPasswordFormTest.php
@@ -1,46 +1,0 @@
-<?php
-
-namespace frontend\tests\unit\forms;
-
-use common\fixtures\UserFixture;
-use frontend\forms\ResetPasswordForm;
-
-class ResetPasswordFormTest extends \Codeception\Test\Unit
-{
-    /**
-     * @var \frontend\tests\UnitTester
-     */
-    protected $tester;
-
-
-    public function _before()
-    {
-        $this->tester->haveFixtures([
-            'user' => [
-                'class' => UserFixture::class,
-                'dataFile' => codecept_data_dir() . 'user.php'
-            ],
-        ]);
-    }
-
-    public function testWrongToken()
-    {
-        $this->tester->expectException('\yii\base\InvalidArgumentException', function() {
-            new ResetPasswordForm('');
-        });
-
-        $this->tester->expectException('\yii\base\InvalidArgumentException', function() {
-            new ResetPasswordForm('notexistingtoken_1391882543');
-        });
-    }
-
-    public function testCorrectToken()
-    {
-        $user = $this->tester->grabFixture('user', 0);
-        print_r($user);
-        $form = new ResetPasswordForm($user['password_reset_token']);
-        $form->password = 'new-password';
-        expect_that($form->validate());
-    }
-
-}
