+++ frontend/views/site/signup.php
@@ -2,7 +2,7 @@
 
 /* @var $this yii\web\View */
 /* @var $form yii\bootstrap\ActiveForm */
-/* @var $model \frontend\forms\SignupForm */
+/* @var $model \shop\forms\auth\SignupForm */
 
 use yii\helpers\Html;
 use yii\bootstrap\ActiveForm;
