+++ shop/forms/auth/ResetPasswordForm.php
@@ -1,5 +1,5 @@
 <?php
-namespace shop\forms;
+namespace shop\forms\auth;
 
 use yii\base\Model;
